Cleanup: whole eventlistener in project / state
