+++ src/projects/automaton/AutomatonProject.ts
@@ -1,5 +1,5 @@
 import { Project } from '../Project'
-import { computed, onMounted } from 'vue'
+import { computed, onMounted, onBeforeUnmount } from 'vue'
 import { stateManager } from '@/projects/stateManager'
 import { registerProjectType } from '../projectRegistry'
 import AutomatonPropsComponent from './AutomatonPropsComponent.vue'
@@ -7,6 +7,31 @@ import type { AutomatonProps, AutomatonState } from './AutomatonTypes'
 import { createPanel } from '@/utility/dockview/integration'
 
 export type { AutomatonProps, AutomatonState } from './AutomatonTypes'
+
+// listener for export fsm -> statetable
+export function useFsmListener() {
+    const handler = (event: MessageEvent) => {
+        if (event.data?.action === 'export') {
+            console.log('raw data:', event.data.fsm)
+
+            const fsmData: AutomatonState = {
+                states: event.data.fsm.states || [],
+                transitions: event.data.fsm.transitions || [],
+                automatonType: event.data.fsm.automatonType
+            }
+            console.log('fsmlistener works. fsmData:', fsmData)
+            stateManager.state.automaton = fsmData
+        }
+    }
+
+    onMounted(() => {
+        window.addEventListener('message', handler)
+        console.log('fsm event listener mounted')
+    })
+    onBeforeUnmount(() => {
+        window.removeEventListener('message', handler)
+    })
+}
 
 // reactive postmessage listener
 const fsmHandler = (event: MessageEvent) => {
